requirements.txt
fastmcp
paramiko
python-dotenv
